% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{bad_arg}
\alias{bad_arg}
\title{Print an error to console (generic \code{ggDNAvis} helper)}
\usage{
bad_arg(
  argument_name,
  arguments_list,
  message,
  class = "argument_value_or_type"
)
}
\arguments{
\item{argument_name}{\code{character}. The name of the argument that caused the error}

\item{arguments_list}{\code{list}. A named list where \code{arguments_list[[argument_name]]} is the value of the offending argument.}

\item{message}{\code{character}. The message that should be printed to describe why the argument is invalid.}

\item{class}{\code{character}. The class that the error should have. Defaults to \code{"argument_value_or_type"} for my own use.}
}
\value{
Nothing, but causes an error exit via \code{\link[rlang:abort]{rlang::abort()}}
}
\description{
This function takes an argument name, a named list of arguments
(presumably being iterated over for a particular validation check),
and a message. It prints an error message of the form:
Argument '<argument_name>' \if{html}{\out{<message>}}
Current value: <argument_value>
}
\examples{
## Obviously this error message function causes
## an error, so needs to be wrapped in try()
positive_args <- list(number = -1)
try(bad_arg("number", positive_args, "must be positive"))

}
