% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_methylation.R
\name{convert_probabilities_to_annotations}
\alias{convert_probabilities_to_annotations}
\title{Create dataframe of locations and rendered probabilities (\code{\link[=visualise_methylation]{visualise_methylation()}} helper)}
\usage{
convert_probabilities_to_annotations(
  modification_locations,
  modification_probabilities,
  sequences,
  sequence_text_scaling = c(-0.5, 256),
  sequence_text_rounding = 2
)
}
\arguments{
\item{modification_locations}{\verb{character vector}. One character value for each sequence, storing a condensed string (e.g. \code{"3,6,9,12"}, produced via \code{\link[=vector_to_string]{vector_to_string()}}) of the indices along the read at which modification was assessed. Indexing starts at 1.}

\item{modification_probabilities}{\verb{character vector}. One character value for each sequence, storing a condensed string (e.g. \code{"0,128,255,15"}, produced via \code{\link[=vector_to_string]{vector_to_string()}}) of the probability of methylation/modification at each assessed base.\cr\cr Assumed to be Nanopore > SAM style modification stored as an 8-bit integer from 0 to 255, but changing other arguments could make this work on other scales.}

\item{sequences}{\verb{character vector}. One character value for each sequence, storing the actual DNA sequence.}

\item{sequence_text_scaling}{\verb{numeric vector, length 2}. The min and max possible probability values, used to facilitate scaling of the text in each to 0-1. Scaling is implemented as \eqn{\frac{p - min}{max}}, so custom scalings (e.g. scaled to 0-9 space) can be implemented by setting this values as required.\cr\cr Set to \code{c(0, 1)} to not scale at all i.e. print the raw integer probability values. It is recommended to also set \code{sequence_text_rounding = 0} to print integers as the default value of \code{2} will result in e.g. \code{"128.00"}. \cr\cr Set to \code{c(-0.5, 256)} (default, results in \eqn{\frac{p+0.5}{256}}) to scale to the centre of the probability spaces defined by the SAMtools spec, where integer \eqn{p} represents the probability space from \eqn{\frac{p}{256}} to \eqn{\frac{p+1}{256}}. This is slightly better at representing the uncertainty compared to \code{c(0, 255)} as strictly speaking \code{0} represents the probability space from 0.000 to 0.004 and \code{255} represents the probability space from 0.996 to 1.000, so scaling them to 0.002 and 0.998 respectively is a more accurate representation of the probability space they each represent. Setting \code{c(0, 255)} would scale such that 0 is exactly 0.000 and 255 is exactly 1.000, which is not as accurate so it discouraged.}

\item{sequence_text_rounding}{\code{integer}. How many decimal places the text drawn in the boxes should be rounded to (defaults to \code{2}). Ignored if \code{sequence_text_type} is \code{"sequence"} or \code{"none"}.}
}
\value{
\code{dataframe}. Dataframe of \code{x_position}, \code{y_position}, \code{annotation} (i.e. probability to draw), and \code{type} (always \code{"Probability"}).
}
\description{
This function takes the locations/probabilities/sequences input to \code{\link[=visualise_methylation]{visualise_methylation()}},
as well as the scaling and rounding to apply to the probability text,
and produces a dataframe of the x and y coordinates to draw each probability at
(i.e. inside the coloured box for each assessed base)
and the probability text to draw inside each box.
}
\examples{
d <- extract_methylation_from_dataframe(example_many_sequences)

## Unscaled i.e. integers
convert_probabilities_to_annotations(
    d$locations,
    d$probabilities,
    d$sequences,
    sequence_text_scaling = c(0, 1),
    sequence_text_rounding = 0
)

## Scaled to 0-1, 3 dp
convert_probabilities_to_annotations(
    d$locations,
    d$probabilities,
    d$sequences,
    sequence_text_scaling = c(-0.5, 256),
    sequence_text_rounding = 3
)

## Default (i.e. scaled to 0-1, 2 dp)
convert_probabilities_to_annotations(
    d$locations,
    d$probabilities,
    d$sequences
)

}
