% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{convert_sequences_to_matrix}
\alias{convert_sequences_to_matrix}
\title{Convert vector of sequences to character matrix (generic \code{ggDNAvis} helper)}
\usage{
convert_sequences_to_matrix(sequences, line_length = NA, blank_value = NA)
}
\arguments{
\item{sequences}{\verb{character vector}. The sequences to transform into a matrix}

\item{line_length}{\code{integer}. The width of the matrix. Set to \code{NA} (default) to automatically use the length of the longest sequence in \code{sequences}.}

\item{blank_value}{\code{value}. The value that should be used to fill in blank/missing points of the matrix.}
}
\value{
\code{matrix}. A matrix of the sequences with one line per sequence, ready for rasterisation via \code{\link[=rasterise_matrix]{rasterise_matrix()}}.
}
\description{
This function takes a vector of sequences (e.g. input to \code{\link[=visualise_many_sequences]{visualise_many_sequences()}}
or \code{\link[=visualise_methylation]{visualise_methylation()}}, or vector split from input to \code{\link[=visualise_single_sequence]{visualise_single_sequence()}}).
It converts it into a matrix e.g. \code{c("GGCGGC", "", "ACGT", "")} would become:
\preformatted{G  G  C  G  G  C
NA NA NA NA NA NA
A  C  G  T  NA NA
NA NA NA NA NA NA
}

The resulting matrix can then be rasterised into a coordinate-value dataframe via \code{\link[=rasterise_matrix]{rasterise_matrix()}}.
}
\examples{
convert_sequences_to_matrix(
    sequences = c("GGCGGC", "", "ACGT", "")
)

convert_sequences_to_matrix(
    sequences = c("GGCGGC", "", "ACGT", ""),
    line_length = 10,
    blank_value = "X"
)

}
