% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_methylation.R
\name{create_plot_masks}
\alias{create_plot_masks}
\title{Create a masking layer for methylation visualisation (\link{visualise_methylation} helper)}
\usage{
create_plot_masks(
  modification_locations_str,
  modification_probabilities_str,
  max_length,
  sequence_length
)
}
\arguments{
\item{modification_locations_str}{\code{character}. A comma-separated string representing a condensed numerical vector (e.g. \code{"3,6,9,12"}, produced via \code{\link[=vector_to_string]{vector_to_string()}}) of the indices along the read at which modification was assessed. Indexing starts at 1.}

\item{max_length}{\code{integer}. How long the output vector should be.}

\item{sequence_length}{\code{integer}. How long the sequence itself is. If smaller than max_length, the remaining spaces will be filled with \code{0}s i.e. set to the background colour in \code{\link[=visualise_methylation]{visualise_methylation()}}.}

\item{modification_probabilites_str}{\code{character}. A comma-separated string representing a condensed numerical vector (e.g. \code{"2,212,128,64"}, produced via \code{\link[=vector_to_string]{vector_to_string()}}) of the probability of modification as an 8-bit (0-255) integer for each base where modification was assessed.}
}
\value{
\verb{numeric vector}. Vector of length \code{max_length} containing \code{1}s for non-assessed bases (e.g. non-CpG), \code{2}s for assessed bases (e.g. CpG), and \code{0} for squares that aren't bases (i.e. past the end of the sequence when \code{max_length > sequence_length}).
}
\description{
Takes modification locations (indices along the read signifying bases at which
modification probability was assessed) and modification probabilities (the probability
of modification at each assessed location, as an integer from 0 to 255), as comma-separated
strings (e.g. \code{"1,5,25"}) produced from numerical vectors via \code{\link[=vector_to_string]{vector_to_string()}}. \cr\cr
Also takes the length of this particular sequence, and the maximum length of sequences in the
collection being methylation-visualised.\cr\cr
Returns a numeric vector with value \code{0} for background bases (past the end of the sequence),
\code{1} for non-assessed bases (e.g. non-CpG for most methylation callers), and \code{2} for assessed bases
(e.g. CpG). This allows controlling the appearance of these categories independently.
}
