% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{format_time_diff}
\alias{format_time_diff}
\title{Format a difference between times (generic \code{ggDNAvis} helper)}
\usage{
format_time_diff(new_time, old_time, characters_to_print = 4)
}
\arguments{
\item{new_time}{\code{POSIXct}. The more recent (newer) of the two times to calculate a difference between.}

\item{old_time}{\code{POSIXct}. The less recent (older) of the two times to calculate a difference between.}

\item{characters_to_print}{\code{integer}. How many numeric digits should be printed.}
}
\value{
\code{character}. The formatted time difference in seconds.
}
\description{
This function takes two times (class \code{"POSIXct"}) and formats
the difference between them nicely, with a certain number
of numerical characters printed.

Note that the if the time difference rounded to the integer
number of seconds (e.g. 1234 seconds) requires more space than
the number of characters allocated (e.g. 3 characters) then
it will go beyond the specified characters.
However, this would be an exceptionally slow-running function.
In normal monitoring use for \code{\link[=monitor]{monitor()}},
<1 second steps should be nearly universal, and <0.01 second
steps are very common.
}
\examples{
## POSIXct time is a very large number of seconds
newer <- 1000000001
older <- 1000000000
format_time_diff(newer, older, 4)

newer <- 1000000456.45645
older <- 1000000000
format_time_diff(newer, older, 4)
format_time_diff(newer, older, 3)
format_time_diff(newer, older, 2)

newer <- 1000000000.011
older <- 1000000000
format_time_diff(newer, older, 4)
format_time_diff(newer, older, 3)
format_time_diff(newer, older, 2)

}
