% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDNAvis-package.R
\docType{package}
\name{ggDNAvis-package}
\alias{ggDNAvis}
\alias{ggDNAvis-package}
\title{ggDNAvis: ggplot2-Based Tools For Visualising DNA Sequences And Modifications}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Evelyn Jade \email{evelynjade42@gmail.com}

}
\keyword{internal}
