% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDNAvis-package.R
\name{ggDNAvis_aliases}
\alias{ggDNAvis_aliases}
\alias{aliases}
\title{\code{ggDNAvis} aliases}
\description{
As of v1.0.0, \code{ggDNAvis} supports function and argument aliases.
The code is entirely written with British spellings (e.g. \code{\link[=visualise_methylation_colour_scale]{visualise_methylation_colour_scale()}}),
but should also accept American spellings (e.g. \code{\link[=visualize_methylation_color_scale]{visualize_methylation_color_scale()}}).
If any American spellings don't work, I most likely overlooked them and can easily fix,
so please submit a bug report by creating a github issue
(\url{https://github.com/ejade42/ggDNAvis/issues}).

All four major \code{visualise_} functions have aliases to also accept \code{visualize_}:
\itemize{
\item \code{\link[=visualise_many_sequences]{visualise_many_sequences()}} (\code{\link[=visualize_many_sequences]{visualize_many_sequences()}})
\item \code{\link[=visualise_methylation]{visualise_methylation()}} (\code{\link[=visualize_methylation]{visualize_methylation()}})
\item \code{\link[=visualise_methylation_colour_scale]{visualise_methylation_colour_scale()}} (\code{\link[=visualize_methylation_color_scale]{visualize_methylation_color_scale()}})
\item \code{\link[=visualise_single_sequence]{visualise_single_sequence()}} (\code{\link[=visualize_single_sequence]{visualize_single_sequence()}})
}

As of v1.0.0, \code{extract_methylation_from_dataframe()} has been renamed \code{extract_and_sort_methylation()}
for consistency with \code{extract_and_sort_sequences()}. To preserve compatibility and ensure consistency,
both functions now accept either name formulation:
\itemize{
\item \code{\link[=extract_and_sort_sequences]{extract_and_sort_sequences()}} (\code{\link[=extract_sequences_from_dataframe]{extract_sequences_from_dataframe()}})
\item \code{\link[=extract_and_sort_methylation]{extract_and_sort_methylation()}} (\code{\link[=extract_methylation_from_dataframe]{extract_methylation_from_dataframe()}})
}

The builtin dataset \code{\link{sequence_colour_palettes}}, like all \code{colour} arguments, also accepts
\code{color} or \code{col}:
\itemize{
\item \code{\link{sequence_colour_palettes}} (\code{\link{sequence_color_palettes}} & \code{\link{sequence_col_palettes}})
}

Additionally, the three \code{rasterise_} helper functions also accept \code{rasterize_}:
\itemize{
\item \code{\link[=rasterise_matrix]{rasterise_matrix()}} (\code{\link[=rasterize_matrix]{rasterize_matrix()}})
\item \code{\link[=rasterise_index_annotations]{rasterise_index_annotations()}} (\code{\link[=rasterize_index_annotations]{rasterize_index_annotations()}})
\item \code{\link[=rasterise_probabilities]{rasterise_probabilities()}} (\code{\link[=rasterize_probabilities]{rasterize_probabilities()}})
}

All arguments should have aliases configured. In particular, any \verb{_colour} arguments
should also accept \verb{_color} or \verb{_col}.

When more than one equivalent argument is provided, the 'canonical' (British) argument
takes precedence, and will produce a warning message explaining this. For colours, \verb{_colour}
takes precedence over \verb{_color}, which itself takes precedence over \verb{_col}.

I have also tried to provide aliases for common argument misspellings. In particular,
\code{index_annotation_full_line} also accepts any of \code{index_annotations_full_lines},
\code{index_annotation_full_lines}, or \code{index_annotations_full_line}.
Likewise, \code{index_annotations_above} also accepts \code{index_annotation_above}.
}
\examples{
d <- extract_methylation_from_dataframe(example_many_sequences)
## The resulting low colour will be green
visualise_methylation(
    d$locations,
    d$probabilities,
    d$sequences,
    index_annotation_lines = NA,
    outline_linewidth = 0,
    high_colour = "white",
    low_colour = "green",
    low_color = "orange",
    low_col = "purple"
)

## The resulting low colour will be orange
visualise_methylation(
    d$locations,
    d$probabilities,
    d$sequences,
    index_annotation_lines = NA,
    outline_linewidth = 0,
    high_colour = "white",
    low_color = "orange",
    low_col = "purple"
)

## The resulting low colour will be purple
visualise_methylation(
    d$locations,
    d$probabilities,
    d$sequences,
    index_annotation_lines = NA,
    outline_linewidth = 0,
    high_colour = "white",
    low_col = "purple"
)

}
