% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_many_sequences.R
\name{insert_at_indices}
\alias{insert_at_indices}
\title{Insert blank items at specified indices in a vector (\code{\link[=visualise_many_sequences]{visualise_many_sequences()}} helper)}
\usage{
insert_at_indices(
  original_vector,
  insertion_indices,
  insert_before = TRUE,
  insert = ""
)
}
\arguments{
\item{original_vector}{\code{vector}. The vector to insert blanks into at specified locations (e.g. vector of sequences from extract_and_sort_sequences, but doesn't have to be).}

\item{insertion_indices}{\verb{integer vector}. The indices (1-indexed) at which blanks should be inserted. If length 0, no blanks will be inserted.}

\item{insert_before}{\code{logical}. Whether blanks should be inserted before (\code{TRUE}, default) or after (\code{FALSE}) each specified index.}

\item{insert}{\code{value}. The value that should be inserted before/after each specified index. Defaults to \code{""}. If length 0, nothing will be inserted. If length > 1, multiple items will be inserted at each specified index.}
}
\value{
\code{vector}. The original vector but with the \code{insert} value added before/after each specified index.
}
\description{
This function takes a vector (e.g. the output of \code{\link[=extract_and_sort_sequences]{extract_and_sort_sequences()}}) and
inserts a specified "blank" value at the specified indices.
If \code{insert_before} is \code{TRUE} then the blank value will be inserted before each
specified index, whereas if \code{insert_before} is \code{FALSE} then the blank value
will be inserted after each specified index.
}
\examples{
insert_at_indices(c("A", "B", "C", "D", "E"), c(2, 4))

insert_at_indices(
    c("A", "B", "C", "D", "E"),
    c(4, 2),
    insert_before = TRUE,
    insert = 0
)

insert_at_indices(
    c("A", "B", "C", "D", "E"),
    c(4, 2),
    insert_before = FALSE,
    insert = 0
)

insert_at_indices(
    original_vector = c("A", "B", "C", "D", "E"),
    insertion_indices = c(1, 4, 6),
    insert_before = TRUE,
    insert = c("X", "Y")
)

insert_at_indices(
    list("A", "B", "C", "D", "E"),
    c(2, 4),
    insert = TRUE
)

insert_at_indices(
    list("A", "B", "C", "D", "E"),
    c(2, 4),
    insert_before = FALSE,
    insert = list(TRUE, 7)
)

insert_at_indices(
    NA,
    c(1, 2),
    FALSE
)

}
