% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{monitor}
\alias{monitor}
\title{Continue performance monitoring (generic \code{ggDNAvis} helper)}
\usage{
monitor(monitor_performance, start_time, previous_time, message)
}
\arguments{
\item{monitor_performance}{\code{logical}. Whether verbose performance monitoring should be enabled.}

\item{start_time}{\code{POSIXct}. The time at which the overarching function was initialised (generally via \code{\link[=monitor_start]{monitor_start()}}).}

\item{previous_time}{\code{POSIXct}. The time at which the previous step was recorded (via a prior call to \code{\link[=monitor]{monitor()}}).}

\item{message}{\code{character}. The message to be printed, generally indicating what this step is doing}
}
\value{
\code{POSIXct} the time at which the function was called, via \code{\link[=Sys.time]{Sys.time()}}.
}
\description{
This function is meant to be called frequently throughout
a main function, and if verbose performance monitoring is enabled
then it will print the elapsed time since (a) initialisation via
\code{\link[=monitor_start]{monitor_start()}} and (b) since the last step was recorded via
this function.
}
\examples{
## Initialise monitoring
start_time <- monitor_start(TRUE, "my_cool_function")

## Step 1
monitor_time <- monitor(TRUE, start_time, start_time, "performing step 1")
x <- 2 + 2

## Step 2
monitor_time <- monitor(TRUE, start_time, monitor_time, "performing step 2")
y <- 10.5^6 \%\% 345789

## Step 3
monitor_time <- monitor(TRUE, start_time, monitor_time, "performing step 3")
z <- y / x^2

## Conclude monitoring
monitor_time <- monitor(TRUE, start_time, monitor_time, "done")

}
