% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{monitor_start}
\alias{monitor_start}
\title{Start performance monitoring (generic \code{ggDNAvis} helper)}
\usage{
monitor_start(monitor_performance, function_name)
}
\arguments{
\item{monitor_performance}{\code{logical}. Whether verbose performance monitoring should be enabled.}

\item{function_name}{\code{character}. The name of the calling function, printed as part of the monitoring initialisation message.}
}
\value{
\code{POSIXct} the time at which the function was initialised, via \code{\link[=Sys.time]{Sys.time()}}.
}
\description{
This function takes a bool of whether verbose performance
monitoring is on, as well as the name of the calling function,
prints a monitoring initialisation message (if desired),
and returns the start time.

Later monitoring steps are performed by \code{\link[=monitor]{monitor()}}
}
\examples{
## Initialise monitoring
start_time <- monitor_start(TRUE, "my_cool_function")

## Step 1
monitor_time <- monitor(TRUE, start_time, start_time, "performing step 1")
x <- 2 + 2

## Step 2
monitor_time <- monitor(TRUE, start_time, monitor_time, "performing step 2")
y <- 10.5^6 \%\% 345789

## Step 3
monitor_time <- monitor(TRUE, start_time, monitor_time, "performing step 3")
z <- y / x^2

## Conclude monitoring
monitor_time <- monitor(TRUE, start_time, monitor_time, "done")

}
