% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{rasterise_index_annotations}
\alias{rasterise_index_annotations}
\alias{visualize_index_annotations}
\alias{rasterize_index_annotations}
\title{Process index annotations and rasterise to a x/y/layer dataframe (generic \code{ggDNAvis} helper)}
\usage{
rasterise_index_annotations(
  new_sequences_vector,
  original_sequences_vector,
  index_annotation_lines,
  index_annotation_interval = 15,
  index_annotation_full_line = TRUE,
  index_annotations_above = TRUE,
  index_annotation_vertical_position = 1/3,
  index_annotation_always_first_base = FALSE,
  sum_indices = FALSE,
  spacing = NA,
  offset_start = 0
)
}
\arguments{
\item{new_sequences_vector}{\code{vector}. The output of \code{\link[=insert_at_indices]{insert_at_indices()}} when used with identical arguments. Should be \code{original_sequences_vector} with some additional blank lines inserted.}

\item{original_sequences_vector}{\code{vector}. The vector of sequences used for plotting, that was originally given to \code{\link[=visualise_many_sequences]{visualise_many_sequences()}}/\code{\link[=visualise_methylation]{visualise_methylation()}} or split from the input sequence to \code{\link[=visualise_single_sequence]{visualise_single_sequence()}}. Must also have been used as input to \code{\link[=insert_at_indices]{insert_at_indices()}} to create \code{new_sequences_vector}.}

\item{index_annotation_lines}{\verb{integer vector}. The lines (i.e. elements of \code{sequences_vector}) that should have their base incides annotated. 1-indexed e.g. \code{c(1, 10)} would annotate the first and tenth elements of \code{sequences_vector}.\cr\cr Extra lines are inserted above or below (depending on \code{index_annotations_above}) the selected lines - note that the line numbers come from \code{sequences_vector}, so are unaffected by these insertions.\cr\cr Setting to \code{NA} disables index annotations (and prevents adding additional blank lines). Defaults to \code{c(1)} i.e. first sequence is annotated.}

\item{index_annotation_interval}{\code{integer}. The frequency at which numbers should be placed underneath indicating base index, starting counting from the leftmost base. Defaults to \code{15} (every 15 bases along each row).\cr\cr Setting to \code{0} disables index annotations (and prevents adding additional blank lines).}

\item{index_annotation_full_line}{\code{logical}. Whether index annotations should continue to the end of the longest sequence (\code{TRUE}, default) or should only continue as far as each selected line does (\code{FALSE}).}

\item{index_annotations_above}{\code{logical}. Whether index annotations should go above (\code{TRUE}, default) or below (\code{FALSE}) each line of sequence.}

\item{index_annotation_vertical_position}{\code{numeric}. How far annotation numbers should be rendered above (if \code{index_annotations_above = TRUE}) or below (if \code{index_annotations_above = FALSE}) each base. Defaults to \code{1/3}.\cr\cr Not recommended to change at all. Strongly discouraged to set below 0 or above 1.}

\item{index_annotation_always_first_base}{\code{logical}. Whether to force the first base in each line to always be annotated regardless of whether it occurs at the \code{index_annotation_interval}. Defaults to \code{FALSE}.}

\item{sum_indices}{\code{logical}. Whether indices should be counted separately along each line (\code{FALSE}, default) or summed along all annotated lines (\code{TRUE}). May behave unexpectedly if \code{TRUE} when \code{annotate_full_lines} is also \code{TRUE}.}

\item{spacing}{\code{integer}. The number of blank lines inserted for each index annotation. Set to \code{NA} (default) to infer from \code{annotation_vertical_position}.}

\item{offset_start}{\code{integer}. The number of blank lines not present at the start, that otherwise would be expected based on \code{spacing} or \code{annotation_vertical_position}. Defaults to \code{0}.}
}
\value{
\code{dataframe}. A dataframe with columns \code{x}, \code{y}, and \code{value}, with one observation per annotation number that needs to be drawn onto the ggplot.
}
\description{
\code{rasterize_index_annotations()} is an alias for \code{rasterise_index_annotations()}.

This function is called by
\code{\link[=visualise_many_sequences]{visualise_many_sequences()}}, \code{\link[=visualise_methylation]{visualise_methylation()}}, and \code{\link[=visualise_single_sequence]{visualise_single_sequence()}}
to create the x/y position data for placing the index annotations on the graph.
Its arguments are either intermediate variables produced by the visualisation functions,
or arguments of the visualisation functions directly passed through.\cr\cr
Returns a dataframe with \code{x_position}, \code{y_position}, and \code{value} columns, where the
values are the index annotations.
}
\examples{
## Set up arguments (e.g. from visualise_many_sequences() call)
sequences_data <- example_many_sequences
index_annotation_lines <- c(1, 23, 37)
index_annotation_interval <- 10
index_annotations_above <- TRUE
index_annotation_full_line <- FALSE
index_annotation_vertical_position <- 1/3


## Create sequences vector
sequences <- extract_and_sort_sequences(
    example_many_sequences,
    grouping_levels = c("family" = 8, "individual" = 2)
)
sequences

## Insert blank rows as needed
new_sequences <- insert_at_indices(
    sequences,
    insertion_indices = index_annotation_lines,
    insert_before = index_annotations_above,
    insert = "",
    vert = index_annotation_vertical_position
)
new_sequences

## Create annnotation dataframe
rasterise_index_annotations(
    new_sequences_vector = new_sequences,
    original_sequences_vector = sequences,
    index_annotation_lines = index_annotation_lines,
    index_annotation_interval = 10,
    index_annotation_full_line = index_annotation_full_line,
    index_annotations_above = index_annotations_above,
    index_annotation_vertical_position = index_annotation_vertical_position,
    sum_indices = FALSE,
    spacing = NA, ## infer from vertical position
    offset_start = 0
)


}
