% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{resolve_alias}
\alias{resolve_alias}
\title{Resolve argument value when aliases are used}
\usage{
resolve_alias(
  primary_name,
  primary_val,
  alias_name,
  alias_val,
  primary_default
)
}
\arguments{
\item{primary_name}{\code{character}. The usual name of the argument.}

\item{primary_val}{\code{value}. The value of the argument under its usual name.}

\item{alias_name}{\code{character}. An alternative alias name for the argument.}

\item{alias_val}{\code{value}. The value of the argument under its alias. Expected to be \code{NULL} if the alias argument is not being used.}

\item{primary_default}{\code{value}. The default value of the argument under its usual name, used to determine if the primary argument has been explicitly set.}
}
\value{
\code{value}. Either \code{primary_val} or \code{alias_val}, depending on the logic above.
}
\description{
See the \link{aliases} page for a general explanation of how aliases are used in \code{ggDNAvis}.

This function takes the name and value for the 'primary' form of an argument
(generally British spellings in \code{ggDNAvis}), the name and value of an alternative
'alias' form, and the default value of the 'primary' argument.

If the alias has not been used (i.e. the 'alias' value is \code{NULL}) or if the 'primary'
value has been changed from the default, then the 'primary' value will be returned.
(Note that if the 'alias' value is not \code{NULL} and the 'primary' value has been changed
from the default, then the updated 'primary' value 'wins' and is returned, but with a
warning that explains that both values were set and the 'alias' has been discarded).

If the alias has been used (i.e. the 'alias' value is not \code{NULL}) and the 'primary'
value is the default, then the 'alias' value will be returned.
}
