% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{resolve_alias_map}
\alias{resolve_alias_map}
\title{Process an alias map list (generic \code{ggDNAvis} helper)}
\usage{
resolve_alias_map(alias_map, dots_env, target_env = parent.frame())
}
\arguments{
\item{alias_map}{\code{list}. A list where each entry takes the name of a formal argument in the calling function, and each value is a list containing \code{"default"} (the default value of the formal argument) and \code{"aliases"} (a character vector of all allowed aliases for the formal argument). Aliases are processed in the order given in the character vector, with earlier aliases taking precedence.}

\item{dots_env}{\code{environment}. The environment created from the dots list. \emph{WILL BE MODIFIED} by this function - alias is removed if it exists, to allow searching this environment for any unused arguments.}

\item{target_env}{\code{environment}. The environment in which variables should be modified. Generally \code{\link[=parent.frame]{parent.frame()}} i.e. the calling function.}
}
\value{
Nothing (variables are modified within the \code{target_env}).
}
\description{
See the \link{aliases} page for a general explanation of how aliases are used in \code{ggDNAvis}.

This function takes an alias map and the environment constructed from non-formal
arguments (...) to the calling function, and optionally an environment to function inside,
and works through the aliases provided in the map via \code{\link[=resolve_alias]{resolve_alias()}}.

If any arguments were given that aren't in the alias map an error is raised.
}
\examples{
## Alias map (from within function code)
alias_map <- list(
    low_colour = list(default = "blue", aliases = c("low_color", "low_col")),
    high_colour = list(default = "red", aliases = c("high_color", "high_col"))
)

## Default values (would come from formal arguments)
low_colour = "blue"
high_colour = "red"

## Extra arguments provided by name
dots_env <- list2env(list("low_col" = "black", "low_color" = "white"))

## Process
resolve_alias_map(alias_map, dots_env)

## See values
print(low_colour)
print(high_colour)

}
