% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{string_to_vector}
\alias{string_to_vector}
\title{Split a \code{","}-joined string back to a numerical vector (generic \code{ggDNAvis} helper)}
\usage{
string_to_vector(string)
}
\arguments{
\item{string}{\code{character}. A comma-separated string (e.g. \code{"1,2,3"}) to convert back to a numeric vector.}
}
\value{
\verb{numeric vector}. The same numbers but as a numeric vector (e.g. \code{c(1,2,3)}).
}
\description{
Takes a string (character) produced by \code{\link[=vector_to_string]{vector_to_string()}} and recreates the numeric vector
}
