% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_methylation_from_fastq.R
\name{strip_leading_at}
\alias{strip_leading_at}
\title{Strip leading \code{@} from character vector}
\usage{
strip_leading_at(string)
}
\arguments{
\item{string}{\verb{character vector}. A vector (e.g. read IDs) to strip of a single leading "\code{@}" each wherever one is present.}
}
\value{
\verb{character vector}. The same string but with one "\code{@}" removed from each element that started with one.
}
\description{
This function removes a single leading \code{@} character
from each element of a character vector when present.
This is intended to deal with SAMtools > FASTQ translation
often prefixing read IDs with an "\code{@}", which can result
in read ID mismatches and metadata merging fails.
}
\examples{
strip_leading_at(c("read_1", "@read_2", "@@read_3", "", NA, NULL))
}
