% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_methylation_from_fastq.R
\name{write_modified_fastq}
\alias{write_modified_fastq}
\title{Write modification information stored in dataframe back to modified FASTQ}
\usage{
write_modified_fastq(
  dataframe,
  filename = NA,
  read_id_colname = "read",
  sequence_colname = "sequence",
  quality_colname = NA,
  locations_colnames = c("methylation_locations"),
  probabilities_colnames = c("methylation_probabilities"),
  modification_prefixes = c("C+m?"),
  return = TRUE
)
}
\arguments{
\item{dataframe}{\code{dataframe}. Dataframe containing modification information to write back to modified
UNFINISHED!!!!!}
}
\description{
This function takes a dataframe containing DNA modification information
(e.g. produced by \code{\link[=insert_function_here]{insert_function_here()}}) and writes it back to modified
FASTQ, equivalent to what would be produced via \verb{samtools fastq -T MM,ML}.\cr\cr
Arguments give the names of columns within the dataframe from which to read.\cr\cr
If multiple types of modification have been assessed (e.g. both methylation
and hydroxymethylation), then multiple colnames must be provided for locations
and probabilites, and multiple prefixes (e.g. \code{"C+h?"}) must be provided.\cr\cr
If quality isn't known (e.g. there was a FASTA step at some point in the pipeline),
the \code{quality} argument can be set to \code{NA} to fill in quality scores with \code{"B"}. This
is the same behaviour as \code{SAMtools} v1.21 when converting FASTA to SAM/BAM then FASTQ.
I don't really know why \code{SAMtools} decided the default quality should be "B" but there
was probably a reason so I have stuck with that.
}
