% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{bad_arg}
\alias{bad_arg}
\title{Print an error to console (generic \code{ggDNAvis} helper)}
\usage{
bad_arg(
  argument_name,
  arguments_list,
  message,
  class = "argument_value_or_type",
  force_names = FALSE
)
}
\arguments{
\item{argument_name}{\code{character}. The name of the argument that caused the error}

\item{arguments_list}{\code{list}. A named list where \code{arguments_list[[argument_name]]} is the value of the offending argument.}

\item{message}{\code{character}. The message that should be printed to describe why the argument is invalid.}

\item{class}{\code{character}. The class that the error should have. Defaults to \code{"argument_value_or_type"} for my own use.}

\item{force_names}{\code{logical}. Whether the names argument should be printed even if names is \code{NULL}. Defaults to \code{FALSE}.}
}
\value{
Nothing, but causes an error exit via \code{\link[rlang:abort]{rlang::abort()}}
}
\description{
This function takes an argument name, a named list of arguments
(presumably being iterated over for a particular validation check),
and a message. It prints an error message of the form:
\preformatted{Argument '<argument_name>' <message>
Current value: <argument_value>
}
If the argument value is a named item (i.e. \code{names(arguments_list[[argument_name]])}
is not null), or if \code{force_names} is \code{TRUE}, then the form will be:
\preformatted{Argument '<argument_name>' <message>
Current value: <argument_value>
Current names: <argument_names>
}
}
\examples{
## Obviously this error-message function causes an error,
## so needs to be wrapped in try() for these examples

## Standard use
positive_args <- list(number = -1)
try(bad_arg("number", positive_args, "must be positive"))

## Automatically detects named item and prints names
named <- list(x = c("first item" = 1, "second item" = 7))
try(bad_arg("x", named, "is not acceptable"))

## Can force name printing
try(bad_arg("number", positive_args, "must be positive", force_names = TRUE))

}
