% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDNAvis-package.R
\docType{data}
\name{example_many_sequences}
\alias{example_many_sequences}
\title{Example multiple sequences data}
\format{
\subsection{\code{example_many_sequences}}{

A dataframe with 23 rows and 5 columns:
\describe{
\item{family}{Participant family}
\item{individual}{Participant ID}
\item{read}{Unique read ID}
\item{sequence}{DNA sequence of the read}
\item{sequence_length}{Length (nucleotides) of the read}
}
}
}
\usage{
example_many_sequences
}
\description{
A collection of made-up sequences in the style of long reads over a repeat region
(e.g. \emph{NOTCH2NLC}), with meta-data describing the participant each read is from and
the family each participant is from. Can be used in \code{visualise_many_sequences()} and
helper functions to visualise these sequences.
}
\keyword{datasets}
