% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDNAvis-package.R
\docType{package}
\name{ggDNAvis-package}
\alias{ggDNAvis}
\alias{ggDNAvis-package}
\title{ggDNAvis: ggplot2-Based Tools For Visualising DNA Sequences And Modifications}
\description{
Uses ggplot2 to visualise either (a) a single DNA/RNA sequence split across multiple lines, (b) multiple DNA/RNA sequences, each occupying a whole line, or (c) base modifications such as DNA methylation called by modified bases models in Dorado or Guppy. Functions starting with `visualise_<something>()` are the main plotting functions, and functions starting with `extract_<something>()` are key helper functions for reading files and reformatting data. A full manual is available at https://github.com/ejade42/ggDNAvis
}
\author{
\strong{Maintainer}: Evelyn Jade \email{evelynjade42@gmail.com}

}
\keyword{internal}
