% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDNAvis-package.R
\docType{package}
\name{ggDNAvis-package}
\alias{ggDNAvis}
\alias{ggDNAvis-package}
\title{ggDNAvis: 'ggplot2'-Based Tools for Visualising DNA Sequences and Modifications}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Uses 'ggplot2' to visualise either (a) a single DNA/RNA sequence split across multiple lines, (b) multiple DNA/RNA sequences, each occupying a whole line, or (c) base modifications such as DNA methylation called by modified bases models in Dorado or Guppy. Functions starting with visualise_<something>() are the main plotting functions, and functions starting with extract_<something>() are key helper functions for reading files and reformatting data. Source code is available at \url{https://github.com/ejade42/ggDNAvis} and a full non-expert user guide is available at \url{https://ejade42.github.io/ggDNAvis/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ejade42.github.io/ggDNAvis/}
  \item \url{https://github.com/ejade42/ggDNAvis}
  \item Report bugs at \url{https://github.com/ejade42/ggDNAvis/issues}
}

}
\author{
\strong{Maintainer}: Evelyn Jade \email{evelynjade42@gmail.com} (\href{https://orcid.org/0009-0003-7761-5425}{ORCID}) [copyright holder]

}
\keyword{internal}
