% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDNAvis-package.R
\name{ggDNAvis_shinyapp}
\alias{ggDNAvis_shinyapp}
\alias{ggDNAvis_shiny}
\title{Run the interactive \code{ggDNAvis} shinyapp}
\usage{
ggDNAvis_shinyapp()
}
\description{
\code{ggDNAvis_shiny()} is an alias for \code{ggDNAvis_shinyapp()} - see \link{aliases}.

The \code{ggDNAvis} shinyapp is an interactive frontend for the \code{ggDNAvis} functions.
Arguments can be configured via text/numerical/colour/checkbox entry rather than
on the command line. In the future it will be hosted online, but is
currently accessible only by running the shinyapp locally.

This function checks 'suggests' packages are present
(not needed for main package, but needed for the shinyapp)
and then runs the shinyapp in the \code{inst/shinyapp} directory.
}
\examples{
\dontrun{
ggDNAvis_shinyapp()
}
}
