% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_tools.R
\name{rasterise_index_annotations}
\alias{rasterise_index_annotations}
\alias{rasterize_index_annotations}
\title{Process index annotations and rasterise to a x/y/layer dataframe (generic \code{ggDNAvis} helper)}
\usage{
rasterise_index_annotations(
  new_sequences_vector,
  original_sequences_vector,
  original_indices_to_annotate,
  annotation_interval = 15,
  annotate_full_lines = TRUE,
  annotations_above = TRUE,
  annotation_vertical_position = 1/3,
  sum_indices = FALSE,
  spacing = NA,
  offset_start = 0
)
}
\arguments{
\item{new_sequences_vector}{\code{vector}. The output of \code{\link[=insert_at_indices]{insert_at_indices()}} when used with identical arguments.}

\item{original_sequences_vector}{\code{vector}. The vector of sequences used for plotting, that was originally given to \code{\link[=visualise_many_sequences]{visualise_many_sequences()}}/\code{\link[=visualise_methylation]{visualise_methylation()}} or split from the input sequence to \code{\link[=visualise_single_sequence]{visualise_single_sequence()}}. Must also have been used as input to \code{\link[=insert_at_indices]{insert_at_indices()}} to create \code{new_sequences_vector}.}

\item{original_indices_to_annotate}{\verb{positive integer vector}. The vector of lines (i.e. indices) of \code{original_vector} to be annotated. Read from \code{index_annotation_lines} argument to \code{\link[=visualise_many_sequences]{visualise_many_sequences()}}/\code{\link[=visualise_methylation]{visualise_methylation()}} (but after processing, so is assumed to be unique and sorted). Must also have been used as input to \code{\link[=insert_at_indices]{insert_at_indices()}} to create \code{new_sequences_vector}. Setting to a length-0 value (e.g. \code{numeric(0)}) causes this function to return an empty dataframe.}

\item{annotation_interval}{\code{integer}. The frequency at which numbers should be placed underneath indicating base index, starting counting from the leftmost base. Setting to \code{0} causes this function to return an empty dataframe. Defaults to \code{15}.}

\item{annotate_full_lines}{\code{logical}. Whether annotations should be calculated up to the end of the longest line (\code{TRUE}, default) or to the end of each line being annotated (\code{FALSE}).}

\item{annotations_above}{\code{logical}. Whether annotations should be drawn above (\code{TRUE}, default) or below (\code{FALSE}) each annotated line. Must also have been used as input to \code{\link[=insert_at_indices]{insert_at_indices()}} to create \code{new_sequences_vector}.}

\item{annotation_vertical_position}{\code{numeric}. The vertical position above/below each annotated line that annotations should be drawn. Must also have been used as input to \code{\link[=insert_at_indices]{insert_at_indices()}} to create \code{new_sequences_vector}.}

\item{sum_indices}{\code{logical}. Whether indices should be counted separately along each line (\code{FALSE}, default) or summed along all annotated lines (\code{TRUE}). May behave unexpectedly if \code{TRUE} when \code{annotate_full_lines} is also \code{TRUE}.}

\item{spacing}{\code{integer}. The number of blank lines inserted for each index annotation. Set to \code{NA} (default) to infer from \code{annotation_vertical_position}.}

\item{offset_start}{\code{integer}. The number of blank lines not present at the start, that otherwise would be expected based on \code{spacing} or \code{annotation_vertical_position}. Defaults to \code{0}.}
}
\value{
\code{dataframe}. A dataframe with columns \code{x}, \code{y}, and \code{value}, with one observation per annotation number that needs to be drawn onto the ggplot.
}
\description{
This function is called by
\code{\link[=visualise_many_sequences]{visualise_many_sequences()}}, \code{\link[=visualise_methylation]{visualise_methylation()}}, and \code{\link[=visualise_single_sequence]{visualise_single_sequence()}}
to create the x/y position data for placing the index annotations on the graph.
Its arguments are either intermediate variables produced by the visualisation functions,
or arguments of the visualisation functions directly passed through.\cr\cr
Returns a dataframe with \code{x_position}, \code{y_position}, and \code{value} columns, where the
values are the index annotations.
}
\examples{
## Set up arguments (e.g. from visualise_many_sequences() call)
sequences_data <- example_many_sequences
index_annotation_lines <- c(1, 23, 37)
index_annotation_interval <- 10
index_annotations_above <- TRUE
index_annotation_full_line <- FALSE
index_annotation_vertical_position <- 1/3


## Create sequences vector
sequences <- extract_and_sort_sequences(
    example_many_sequences,
    grouping_levels = c("family" = 8, "individual" = 2)
)
sequences

## Insert blank rows as needed
new_sequences <- insert_at_indices(
    sequences,
    insertion_indices = index_annotation_lines,
    insert_before = index_annotations_above,
    insert = "",
    vert = index_annotation_vertical_position
)
new_sequences

## Create annnotation dataframe
rasterise_index_annotations(
    new_sequences_vector = new_sequences,
    original_sequences_vector = sequences,
    original_indices_to_annotate = index_annotation_lines,
    annotation_interval = 10,
    annotate_full_lines = index_annotation_full_line,
    annotations_above = index_annotations_above,
    annotation_vertical_position = index_annotation_vertical_position,
    sum_indices = FALSE,
    spacing = NA ## infer from vertical position,
    offset_start = 0
)


}
