% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_methylation_from_fastq.R
\name{write_modified_fastq}
\alias{write_modified_fastq}
\title{Write modification information stored in dataframe back to modified FASTQ}
\usage{
write_modified_fastq(
  dataframe,
  filename = NA,
  read_id_colname = "read",
  sequence_colname = "sequence",
  quality_colname = NA,
  locations_colnames = c("hydroxymethylation_locations", "methylation_locations"),
  probabilities_colnames = c("hydroxymethylation_probabilities",
    "methylation_probabilities"),
  modification_prefixes = c("C+h?", "C+m?"),
  return = TRUE
)
}
\arguments{
\item{dataframe}{\code{dataframe}. Dataframe containing modification information to write back to modified FASTQ. Must have columns for unique read ID, DNA sequence, and at least one set of locations and probabilities for a particular modification type (e.g. 5C methylation).}

\item{filename}{\code{character}. File to write the modified FASTQ to. Recommended to end with \code{.fastq} (warns but works if not). If set to \code{NA} (default), no file will be output, which may be useful for testing/debugging.}

\item{read_id_colname}{\code{character}. The name of the column within the dataframe that contains the unique ID for each read. Defaults to \code{"read"}.}

\item{sequence_colname}{\code{character}. The name of the column within the dataframe that contains the DNA sequence for each read. Defaults to \code{"sequence"}.\cr\cr The values within this column must be DNA sequences e.g. \code{"GGCGGC"}.}

\item{quality_colname}{\code{character}. The name of the column within the dataframe that contains the FASTQ quality scores. If scores are not known, can be set to \code{NA} (default) to fill in quality with \code{"B"}.\cr\cr If not \code{NA}, must correspond to a column where the values are the FASTQ quality scores e.g. \code{"105;<<;:998+***,044.-"}.}

\item{locations_colnames}{\verb{character vector}. Vector of the names of all columns within the dataframe that contain modification locations. Defaults to \code{c("hydroxymethylation_locations", "methylation_locations")}.\cr\cr The values within these columns must be comma-separated strings of indices at which modification was assessed, as produced by \code{\link[=vector_to_string]{vector_to_string()}}, e.g. \code{"3,6,9,12"}.}

\item{probabilities_colnames}{\verb{character vector}. Vector of the names of all columns within the dataframe that contain modification probabilities. Defaults to \code{c("hydroxymethylation_probabilities", "methylation_probabilities")}.\cr\cr The values within the columns must be comma-separated strings of modification probabilities, as produced by \code{\link[=vector_to_string]{vector_to_string()}}, e.g. \code{"0,255,128,78"}.}

\item{modification_prefixes}{\verb{character vector}. Vector of the prefixes to be used for the MM tags specifying modification type. These are usually generated by Dorado/Guppy based on the original modified basecalling settings, and more details can be found in the SAM optional tag specifications. Defaults to \code{c("C+h?", "C+m?")}.\cr\cr \code{locations_colnames}, \code{probabilities_colnames}, and \code{modification_prefixes} must all have the same length e.g. 2 if there were 2 modification types assessed.}

\item{return}{\code{logical}. Boolean specifying whether this function should return the FASTQ (as a character vector of each line in the FASTQ), otherwise it will return \code{invisible(NULL)}. Defaults to \code{TRUE}.}
}
\value{
\verb{character vector}. The resulting modified FASTQ file as a character vector of its constituent lines (or \code{invisible(NULL)} if \code{return} is \code{FALSE}). This is probably mostly useful for debugging, as setting \code{filename} within this function directly writes to FASTQ via \code{\link[=writeLines]{writeLines()}}.
}
\description{
This function takes a dataframe containing DNA modification information
(e.g. produced by \code{\link[=insert_function_here]{insert_function_here()}}) and writes it back to modified
FASTQ, equivalent to what would be produced via \verb{samtools fastq -T MM,ML}.\cr\cr
Arguments give the names of columns within the dataframe from which to read.\cr\cr
If multiple types of modification have been assessed (e.g. both methylation
and hydroxymethylation), then multiple colnames must be provided for locations
and probabilites, and multiple prefixes (e.g. \code{"C+h?"}) must be provided.
These three vectors must all be the same length.\cr\cr
If quality isn't known (e.g. there was a FASTA step at some point in the pipeline),
the \code{quality} argument can be set to \code{NA} to fill in quality scores with \code{"B"}. This
is the same behaviour as SAMtools v1.21 when converting FASTA to SAM/BAM then FASTQ.
I don't really know why SAMtools decided the default quality should be "B" but there
was probably a reason so I have stuck with that.\cr\cr
Default arguments are set up to work with the included \code{example_many_sequences} data.
}
